[MODIFY] 기념일 검색 화면 관련 API 연동 / 가족 선택 유지

+++ app/src/main/java/com/example/wooriga/Anniversary.kt
@@ -3,7 +3,7 @@ package com.example.wooriga.model
 import com.google.gson.annotations.SerializedName
 
 data class Anniversary(
-    val anniversaryId: Int? = null,
+    val anniversaryId: Long? = null,
     val familyId: Int,
 
     val date: String,
